Rename method GetSolenoidSwitch to IsPistonAtLimit.
The solenoids are valves that route air to the piston.  The limit switch
detects the position of the piston, not the solenoids.

The new name also accurately declares our intent when invoking the
method.  We're more interested in whether the piston has reached its
limit than whether a switch is open or closed.

STATUS: Builds.

- Justin C