Add methods with explicit names to control piston.
It's more clear when reading code what its intent is when a single
method named PistonAscend is used instead of SetSolenoidAscend(true) +
SetSolenoidDescend(false).

Mistakes are easier to recognize with this naming scheme over the
former.  Especially when students that copy-paste refuse to trace their
code to verify correctness.

STATUS: Builds.

- Justin C