Add test controls for creeper climb sequence
This is the joystick setup previously used to successfully test the
climb sequence.

It returns in its own method so we can drop a function call in
TeleopPeriodic to resume further testing until the driver station
joystick setup is implemented.  After that, maybe this sticks around for
troubleshooting.

STATUS: Builds.

- Justin C