Remove position prereq from ShootCargoForCargoShip.
Discussed with Ryan during testing.  Sometimes the backlash of the
intake causes the position check to fail.

Let the drive team make the judgement call whether to fire and it's on
them if a misfire occurs.

Might add ability later for drive team to adjust intake rotation
manually.

STATUS: Builds.  Tested on testbot.

- Justin C