Rename GetHatch-Pressed button commands to Is-Down.
There was a bit of confusion while troubleshooting do to the old naming
convention, so trying out a name that is more descriptive.

Use Is for boolean returns to differentiate from axis getters that
return doubles.

Use Down instead of Pressed to differentiate edge-triggered getters from
level-triggered getters.  For example, the GetXButton command is
level-triggered, while GetXButtonPressed is edge-triggered, but they're
both true when the user "presses" the button.

STATUS: Builds.

- Robby S, Logan H-D, Justin C