Stage command TakeCargoFromDispenser.
Currently unused, the intent is to double-up commands on the "Cargo
load/shoot" console button.  When no cargo in intake, use the
TakeCargoFromDispenser.  When cargo in intake, use the
RotateCargoForCargoShip.

TakeCargoFromDispenser should rotate to the carry position after ball is
captured, but the intake is already in position, so no-op.

May be unnecessary as operator can activate TakeCargo and
RotateCargoForCargoShip separately.

STATUS: Builds.

- Robby S, Logan H-D, Justin C