Revert use of Scheduler.ResetAll.
Tried to clear existing commands out of the scheduler when robot is
disabled, so commands don't resume when entering teleop or autonomous.

Unfortunately, the default command for the DriveTrain subsystem died,
and our robot was bereft.

Return to running scheduler in DisablePeriodic to at least clear out
cancelled commands.

STATUS: Builds.  Tested on testbot 2019-03-05.

- Justin C