Build, deploy, and print a version file.
Provide a simple means by which to verify our robot is running the
correct code by deploying a version file and printing the contents to
the dashboard console when the robot is disabled.

Also ignore the version file in git.  We don't want builds reporting the
same version across commits.

STATUS: Builds.

- Justin C