Force versionFile build task always out of date.
An attempt to always execute the versionFile build task and regenerate
the version file, even when nothing else in the project is changed.
This is done to ensure the version file never fails to update when
building agaisnt another revision since the version file doesn't
actually depend on anything else in the project.

STATUS: Builds.

- Justin C