Wire operator console to PID hatch commands.
Add three vertical buttons from the operator console.

Connect two buttons to Grab and Ready commands.

Reuse LowerHatch command for third button.  It disables the pid and
controls the hatch motor while the button is pressed, which covers all
the bases for the third button, no need for an extra command.

STATUS: Builds.

- Justin C