Reverse comparison operator for hatch grip threshold.
Range of compbot hatch position sensor is reverse that of potentiometer
on testbot. Reverse threshold comparison operator so that the threshold
is exceeded when the hatch position indicates the hatch is higher.

The optimal fix is to adjust the range and offset of the position sensor
to { range: -5, offset: 5 } to map the sensor values to those of the
testbot sensor.  But the robot is calibrated, so instead of invalidating
those settings, this commit is applied instead.

STATUS: Builds.  Tested on compbot.

- Logan H-D