Fix acceleration function
- Introduce a minimum acceleration to ensure desired velocity setpoint
  is reached.

- Clamp acceleration, so final velocity is not exceeded.

The velocity is modelled with the equations:

  V[i+1] = (V[f] - V[i]) * A_f + V[o]                 (1)
  V[f] = \sum_{i=1}^N{(V[f] - V[o]) * A_f^i} + V[o]   (2)

where A_f is the acceleration factor between [0..1],
  V[o] is initial velocity,
  V[f] is final velocity,
  V[i] is velocity at some moment between [o..f)

Eq (2) is a variation of:

  g(x) = \sum_{i=1}^x{C^i}          (3)

When C is 0.5, g(x) approaches a value of 1.0 for increasing values of
x.  This value of 1.0 is desired as it allows Eq (2) to reach the
desired velocity setpoint.  For values of C < 0.5, g(x) will approach a
value < 1.0 and never achieve the desired velocity setpoint.  Use max
function to transition from exponential sum to linear sum and ensure the
desired velocity setpoint is reached, and use the min function to ensure
the desired velocity setpoint is not exceeded.

  h(x) = min( 1.0, \sum_{i=1}^x{max(C^i, D_{min})} )      (4)

STATUS: Builds.  Not tested.

- Justin C

Signed-off-by: Justin Charette <charetjc@gmail.com>